requirements.txt
fastapi
uvicorn[standard]
livekit-api>=1.1.0
pydantic
python-dotenv